map with a marker and info window, need to figure out how to place an image inside window
